Add docs for crossbow-permissions
