+++ plugins/permissions/src/android/mod.rs
@@ -1,8 +1,6 @@
-mod check_permission;
 mod request_permission;
 mod show_text;
 
-pub use check_permission::*;
 pub use request_permission::*;
 pub use show_text::*;
